+++ src/bt_init.cpp
@@ -22,7 +22,7 @@ SEXP bt_count_(SEXP, SEXP, SEXP);
 SEXP bt_group_agg_(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP);
 SEXP bt_match_mask_(SEXP, SEXP, SEXP, SEXP, SEXP);
 SEXP bt_group_id_(SEXP, SEXP);
-SEXP bt_join_(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP);
+SEXP bt_join_(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP);
 SEXP bt_first_match_(SEXP, SEXP, SEXP, SEXP, SEXP);
 SEXP bt_range_join_(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP);
 SEXP bt_rolling_join_(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP);
@@ -43,7 +43,7 @@ static const R_CallMethodDef CallEntries[] = {
   { "bt_group_agg_",  (DL_FUNC) &bt_group_agg_,  6 },
   { "bt_match_mask_", (DL_FUNC) &bt_match_mask_, 5 },
   { "bt_group_id_",   (DL_FUNC) &bt_group_id_,   2 },
-  { "bt_join_",         (DL_FUNC) &bt_join_,         7 },
+  { "bt_join_",         (DL_FUNC) &bt_join_,         8 },
   { "bt_first_match_",  (DL_FUNC) &bt_first_match_,  5 },
   { "bt_range_join_",   (DL_FUNC) &bt_range_join_,   9 },
   { "bt_rolling_join_", (DL_FUNC) &bt_rolling_join_, 9 },